cout-ai/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.51.2
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.24.1
python-multipart==0.0.6
supabase==2.1.0
python-jose[cryptography]==3.3.0 